Shrinkbench/requirements.txt
pytorch
torchvision
pandas
matplotlib
numpy
tqdm
pandas
